refactor Armor
